% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AEDi-package.R
\docType{package}
\name{AEDi-package}
\alias{AEDi}
\alias{AEDi-package}
\title{AEDi: Pacote de inclusão de fontes de dados, indicadores, e filtros de variáveis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

AED interativa e adição de dados.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rodrigoesborges/AEDi}
  \item Report bugs at \url{https://github.com/rodrigoesborges/AEDi/issues}
}

}
\author{
\strong{Maintainer}: Rodrigo Borges \email{rodrigo@borges.net.br}

}
\keyword{internal}
