% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbdatawriter.R
\name{db_datawrite}
\alias{db_datawrite}
\title{Do some checking, sanitizing and write to backend database
\code{db_datawrite()} does some checking, sanitizing and writes new data to
backend database}
\usage{
db_datawrite(
  metadf,
  datadf,
  construct,
  grp = F,
  engine = "postgresql",
  sanitize = T
)
}
\arguments{
\item{metadf}{dataframe list with  the following expected structure :
[\link{1}] -> df with
orig_name
data_name
data_desc
[\link{2}] -> data.frame with at least
data_class_id
data_freq_id
dataunit_num   - TBI
dataunit_den   - TBI
datasource_id
data_url <- construct ? call?
@param datadf dataframe with the following expected structure:
local <-> local_id or name/code pairable with local table
periodo <- date or string/number in yyyy,yyyy-mm, yyyy-mm-dd
valor <- numeric
@param construct The call that extracted/obtained the data
@param grp Should a datagroup be defined - for tables with subgroups (as sidra)
@param engine Defaults to postgresql, sqlite TBI.}
}
\description{
Do some checking, sanitizing and write to backend database
\code{db_datawrite()} does some checking, sanitizing and writes new data to
backend database
}
