% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controle_execucao.R
\name{fingerprint_aedidb}
\alias{fingerprint_aedidb}
\title{Fingerprint barato do conteudo do aedidb: volumetria + soma + ultimas
escritas. Muda com qualquer INSERT/UPDATE/replace de indicador.}
\usage{
fingerprint_aedidb()
}
\description{
Fingerprint barato do conteudo do aedidb: volumetria + soma + ultimas
escritas. Muda com qualquer INSERT/UPDATE/replace de indicador.
}
\keyword{internal}
