% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executa_atualizacao.R
\name{listar_scripts_coleta}
\alias{listar_scripts_coleta}
\title{Lista os scripts de coleta executaveis (sem .ignore)}
\usage{
listar_scripts_coleta(raiz = .aedi_raiz())
}
\description{
Lista os scripts de coleta executaveis (sem .ignore)
}
