% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_locais_nivel}
\alias{painel_locais_nivel}
\title{Localidades de um nivel territorial (pela largura do geoloc_id) que
possuem dados, rotuladas por nome (municipios ganham sigla da UF)}
\usage{
painel_locais_nivel(con, nivel_id)
}
\description{
Localidades de um nivel territorial (pela largura do geoloc_id) que
possuem dados, rotuladas por nome (municipios ganham sigla da UF)
}
\keyword{internal}
