% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbprepare.R
\name{prepare_db}
\alias{prepare_db}
\title{Prepare app_db}
\usage{
prepare_db(
  tdbname = "aedidb",
  type = "sqlite",
  userdb = "aedi",
  passwddb = "aEd1#man@gR",
  hostdb = "127.0.0.1",
  postgis = TRUE,
  geo = TRUE
)
}
\arguments{
\item{tdbname}{Name of the database that will be created, defaults to aedidb}

\item{type}{type of db backend - defaults to sqlite, alternative pgsql}

\item{userdb}{Name of the user with connection permissions}

\item{passwddb}{userdb password for connecting to tdbname}

\item{hostdb}{host resolvable domain or address}

\item{postgis}{is postgis extension enabled on pgsql database? defaults to TRUE
@importFrom RSQLite SQLite
@importFrom DBI dbExecute dbGetQuery dbSendQuery}
}
\description{
Writes sqlite database on current dir or pgsql tables to aedi_db
if 'pgsql' , user , password and database should be set up
separately on the corresponding PostgreSQL Server
}
