% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controle_execucao.R
\name{controle_fim}
\alias{controle_fim}
\title{Fecha a execucao e atualiza o estado atual do script}
\usage{
controle_fim(
  nome_script,
  hist_id,
  sucesso,
  etapa = "coleta",
  linhas = NA_integer_,
  mensagem = "",
  hash_estado = NA_character_,
  por = Sys.info()[["user"]],
  projeto = "AEDi"
)
}
\arguments{
\item{projeto}{nome do projeto (raiz do lote) dono do script}
}
\description{
Fecha a execucao e atualiza o estado atual do script
}
