% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controle_execucao.R
\name{controle_inicio}
\alias{controle_inicio}
\title{Registra o inicio de uma execucao (retorna id do historico p/ fechar depois)}
\usage{
controle_inicio(nome_script, etapa = "coleta", por = Sys.info()[["user"]])
}
\description{
Registra o inicio de uma execucao (retorna id do historico p/ fechar depois)
}
