% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executa_atualizacao.R
\name{executar_script_coleta}
\alias{executar_script_coleta}
\title{Executa um unico script de coleta com controle de execucao}
\usage{
executar_script_coleta(arquivo, raiz = .aedi_raiz())
}
\description{
Executa um unico script de coleta com controle de execucao
}
