% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controle_execucao.R
\name{ler_controle}
\alias{ler_controle}
\title{Le o controle de um script (ou de todos)}
\usage{
ler_controle(nome_script = NULL)
}
\arguments{
\item{nome_script}{nome do script de coleta (ex.: "objetivo2_1_aedi");
NULL retorna data.frame com todos}
}
\description{
Le o controle de um script (ou de todos)
}
