% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_branding.R
\name{painel_brand_subtitulo}
\alias{painel_brand_subtitulo}
\title{Subtitulo do painel (usa painel_subtitulo)}
\usage{
painel_brand_subtitulo(default = "AEDi — DW de indicadores")
}
\description{
Subtitulo do painel (usa painel_subtitulo)
}
\keyword{internal}
