% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_niveis}
\alias{painel_niveis}
\title{Niveis territoriais disponiveis no DW: apenas os que possuem dados,
com quantidade de localidades distintas}
\usage{
painel_niveis(con)
}
\description{
Niveis territoriais disponiveis no DW: apenas os que possuem dados,
com quantidade de localidades distintas
}
\keyword{internal}
