% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbprepare.R
\name{prepare_db}
\alias{prepare_db}
\title{Prepare app_db}
\arguments{
\item{tdbname}{Name of the database that will be created
@importFrom RSQLite SQLite
@importFrom DBI dbExecute dbGetQuery dbSendQuery}
}
\description{
Writes sqlite database on current dir
}
